add on to previouse commit

+++ src/com/biorecorder/util/series/DoubleRegularSeries.java
@@ -27,7 +27,7 @@ public class DoubleRegularSeries implements DoubleSeries {
     @Override
     public double get(long index) {
         rangeCheck(index);
-        return (double)(startValue + dataInterval * index);
+        return startValue + dataInterval * index;
     }
 
     public double getDataInterval() {
